requirements.txt
annotated-types==0.7.0
asgiref==3.8.1
certifi==2025.1.31
charset-normalizer==3.4.1
Django==5.1.5
django-cors-headers==4.6.0
django-filter==24.3
django-ratelimit==4.1.0
djangorestframework==3.15.2
djangorestframework_simplejwt==5.4.0
drf-yasg==1.21.8
idna==3.10
inflection==0.5.1
maturin==1.8.1
packaging==24.2
pillow==11.1.0
pydantic==2.10.6
pydantic_core==2.27.2
PyJWT==2.10.1
pytz==2025.1
PyYAML==6.0.2
requests==2.32.3
setuptools==75.8.0
sqlparse==0.5.3
typing_extensions==4.12.2
tzdata==2025.1
uritemplate==4.1.1
urllib3==2.3.0
wheel==0.45.1
zarinpal-python-sdk==0.2
python-dotenv
drf-nested-routers==0.94.2
